misc/HtmlTestSuite/results/test-0175.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml"><head><title>CSS Test: @keyframes from the document don't apply in the shadow tree.</title>
<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>
<link href="mailto:emilio@crisal.io" rel="author" title="Emilio Cobos Ãlvarez"/>
<link href="https://drafts.csswg.org/css-scoping/#selectors-data-model" rel="help"/>
<!--
  Beware of https://github.com/w3c/csswg-drafts/issues/1995 potentially, but
  unlikely, changing the expected result of this test.
-->
<style>
@keyframes myanim {
  from { background: red; }
  to { background: red; }
}
</style>
</head><body><div id="host"/>
<script>
test(function() {
  host.attachShadow({ mode: "open" }).innerHTML = `
    &lt;style&gt;
      #in-shadow {
        width: 100px;
        height: 100px;
        background: blue;
        animation: myanim 10s infinite;
      }
    &lt;/style&gt;
    &lt;div id="in-shadow"&gt;&lt;/div&gt;
  `;

  assert_equals(host.shadowRoot.getElementById('in-shadow').getAnimations().length, 0);
}, "@keyframes from the document don't apply in the shadow tree");
</script>
</body></html>
